sched: expand sysctl_sched_features to unsigned long
hulk inclusion
category: other
bugzilla: https://atomgit.com/openeuler/kernel/issues/8702

--------------------------------

To support future feature bits beyond 32, expand the type from unsigned int
to unsigned long for long-term safety.

+++ kernel/sched/core.c
@@ -135,7 +135,7 @@ static void sched_change_qos_group(struct task_struct *tsk, struct task_group *t
  */
 #define SCHED_FEAT(name, enabled)	\
 	(1UL << __SCHED_FEAT_##name) * enabled |
-const_debug unsigned int sysctl_sched_features =
+const_debug unsigned long sysctl_sched_features =
 #include "features.h"
 	0;
 #undef SCHED_FEAT
